challenge-1/agents/v1/requirements.txt
azure-identity
azure-cosmos
azure-ai-agents==1.2.0b5
agent-framework-core==1.0.0b251016
agent-framework-azure-ai==1.0.0b251016
azure-ai-projects==1.0.0b12